+++ src/main/java/com/github/christiangroth/jsonstore/store/JsonStore.java
@@ -16,7 +16,7 @@ import flexjson.JSONSerializer;
  * Represents a JSON store for a concrete class holding zero to many instances. Access is provided using delegate methods to Java built in
  * stream API. You may use flexjson annotations to control conversion from/to JSON.
  * 
- * @author cgroth
+ * @author Christian Groth
  * @param <T>
  *            concrete type stored in this instance
  */