+++ src/test/java/com/github/christiangroth/jsonstore/JsonStoresBuilderTest.java
@@ -1,0 +1,31 @@
+package com.github.christiangroth.jsonstore;
+
+import java.io.File;
+
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+
+import com.google.common.io.Files;
+
+public class JsonStoresBuilderTest {
+	
+	private File tempDir;
+	
+	@Before
+	public void init() {
+		tempDir = Files.createTempDir();
+	}
+	
+	@Test
+	public void transientStores() {
+		JsonStores stores = JsonStores.builder().build();
+		Assert.assertFalse(stores.isPersistent());
+	}
+	
+	@Test
+	public void persistentStores() {
+		JsonStores stores = JsonStores.builder().storage(tempDir, false, false).build();
+		Assert.assertTrue(stores.isPersistent());
+	}
+}